ДЗ 19 спринта(перевел все активити во фрагементы) После добавления NavigationComponent перестают отображаться экраны(не успеваю доделать до дедлайна)
