+++ app/src/main/java/com/practicum/playlistmaker/main/ui/MainActivity.kt
@@ -6,33 +6,29 @@ import android.os.Bundle
 import android.widget.Button
 //import com.practicum.playlistmaker.presentation.ui.media.MediaActivity
 import com.practicum.playlistmaker.R
-import com.practicum.playlistmaker.mediateka.ui.MediatekaActivity
-import com.practicum.playlistmaker.search.ui.SearchActivity
 
-
-import com.practicum.playlistmaker.settings.ui.SettingsActivity
 
 class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
 
-        val searchButton = findViewById<Button>(R.id.search)
-        val mediaButton = findViewById<Button>(R.id.media_library)
-        val settingsButton = findViewById<Button>(R.id.settings)
-
-        searchButton.setOnClickListener {
-            val searchIntent = Intent(this, SearchActivity::class.java)
-            startActivity(searchIntent)
-        }
-        mediaButton.setOnClickListener {
-            val mediaIntent = Intent(this, MediatekaActivity::class.java)
-            startActivity(mediaIntent)
-        }
-        settingsButton.setOnClickListener {
-            val settingsIntent = Intent(this, SettingsActivity::class.java)
-            startActivity(settingsIntent)
-        }
+//        val searchButton = findViewById<Button>(R.id.search)
+//        val mediaButton = findViewById<Button>(R.id.media_library)
+//        val settingsButton = findViewById<Button>(R.id.settings)
+//
+//        searchButton.setOnClickListener {
+//            val searchIntent = Intent(this, SearchActivity::class.java)
+//            startActivity(searchIntent)
+//        }
+//        mediaButton.setOnClickListener {
+//            val mediaIntent = Intent(this, MediatekaActivity::class.java)
+//            startActivity(mediaIntent)
+//        }
+//        settingsButton.setOnClickListener {
+//            val settingsIntent = Intent(this, SettingsActivity::class.java)
+//            startActivity(settingsIntent)
+//        }
 
     }
 }
