+++ app/src/main/java/com/practicum/playlistmaker/mediateka/ui/MediatekaFragment.kt
@@ -1,0 +1,65 @@
+package com.practicum.playlistmaker.mediateka.ui
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.core.os.bundleOf
+import androidx.fragment.app.Fragment
+import com.google.android.material.tabs.TabLayoutMediator
+import com.practicum.playlistmaker.R
+import com.practicum.playlistmaker.databinding.FragmentMediatekaBinding
+
+class MediatekaFragment: Fragment() {
+
+    private lateinit var binding: FragmentMediatekaBinding
+    private lateinit var tabMediator: TabLayoutMediator
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+        binding = FragmentMediatekaBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        val backButton = binding.arrowBack
+
+//        backButton.setOnClickListener {
+//            finish()
+//        }
+
+//        val trackPosition = intent.getStringExtra("poster") ?: ""
+//        val playlistNumber = intent.getIntExtra("id") ?: ""
+
+        binding.viewPager.adapter = MediatekaViewPagerAdapter(
+            fragmentManager = childFragmentManager,
+            lifecycle = lifecycle,
+//            trackPosition = trackPosition,
+//            playlistNumber = playlistNumber,
+        )
+
+        tabMediator = TabLayoutMediator(binding.tabLayout, binding.viewPager) { tab, position ->
+            when(position) {
+                0 -> tab.text = getString(R.string.favorites_tracks)
+                1 -> tab.text = getString(R.string.playlists)
+            }
+        }
+        tabMediator.attach()
+    }
+    override fun onDestroyView() {
+        super.onDestroyView()
+        tabMediator.detach()
+    }
+
+    companion object {
+
+
+        // Тег для использования во FragmentManager
+        const val TAG = "MediatekaFragment"
+
+        fun createArgs(): Bundle =
+            bundleOf()
+
+    }
+
+}
