+++ app/src/main/java/com/practicum/playlistmaker/mediateka/ui/root/RootActivity.kt
@@ -1,0 +1,29 @@
+package com.practicum.playlistmaker.mediateka.ui.root
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import androidx.fragment.app.commit
+import androidx.navigation.fragment.NavHostFragment
+import androidx.navigation.ui.setupWithNavController
+import com.google.android.material.bottomnavigation.BottomNavigationView
+import com.practicum.playlistmaker.R
+import com.practicum.playlistmaker.databinding.ActivityRootBinding
+import com.practicum.playlistmaker.mediateka.ui.MediatekaFragment
+
+class RootActivity : AppCompatActivity() {
+
+    private lateinit var binding: ActivityRootBinding
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_root)
+
+        binding = ActivityRootBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+
+        val navHostFragment = supportFragmentManager.findFragmentById(R.id.rootFragmentContainerView) as NavHostFragment
+        val navController = navHostFragment.navController
+
+        binding.bottomNavigationView.setupWithNavController(navController)
+
+    }
+}
