+++ app/src/main/res/layout/activity_player.xml
@@ -2,6 +2,7 @@
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
+    tools:context=".player.ui.PlayerActivity"
     style="@style/PlayerActivityScreen">
     <androidx.constraintlayout.widget.Guideline
         android:id="@+id/leftBorderForTrackName"
