+++ app/src/main/res/layout/fragment_mediateka.xml
@@ -1,7 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    style="@style/SettingsActivityScreen">
+    xmlns:tools="http://schemas.android.com/tools"
+    tools:context="com.practicum.playlistmaker.mediateka.ui.MediatekaFragment"
+    style="@style/SettingsActivityScreen"
+    >
 
     <LinearLayout
         android:layout_width="match_parent"
@@ -24,6 +27,7 @@
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         style="@style/tabLayoutMediateka"
+        app:tabIndicatorFullWidth = "true"
         />
 
     <androidx.viewpager2.widget.ViewPager2
