+++ app/src/main/res/layout/fragment_search.xml
@@ -2,6 +2,7 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
+    tools:context="com.practicum.playlistmaker.search.ui.SearchFragment"
     android:id="@+id/container"
     style="@style/SettingsActivityScreen"
     >
