+++ app/src/main/res/layout/fragment_settings.xml
@@ -1,6 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    tools:context="com.practicum.playlistmaker.settings.ui.SettingsFragment"
     style="@style/SettingsActivityScreen"
     >
