+++ app/src/main/res/menu/bottom_navigation_menu.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/searchFragment"
+        android:icon="@drawable/search"
+        android:title="@string/title_search" />
+    <item
+        android:id="@+id/mediatekaFragment"
+        android:icon="@drawable/media_library"
+        android:title="@string/title_mediateka" />
+
+    <item
+        android:id="@+id/settingsFragment"
+        android:icon="@drawable/settings"
+        android:title="@string/title_settings" />
+</menu>
